+++ api/serializers.py
@@ -22,22 +22,6 @@ from rest_framework import pagination
 
 
 # Serializers
-class ProfileSerializer(serializers.ModelSerializer):
-    """
-    """
-    #TODO:Need to validate provider/identity membership on id change
-    username = serializers.CharField(read_only=True, source='user.username')
-    email = serializers.CharField(read_only=True, source='user.email')
-    groups = serializers.CharField(read_only=True, source='user.groups.all')
-    is_staff = serializers.BooleanField(source='user.is_staff')
-    is_superuser = serializers.BooleanField(source='user.is_superuser')
-    selected_identity = IdentityRelatedField(source='user.select_identity')
-
-    class Meta:
-        model = UserProfile
-        exclude = ('id',)
-
-
 class ProviderMachineSerializer(serializers.ModelSerializer):
     #R/O Fields first!
     alias = serializers.CharField(read_only=True, source='identifier')
