+++ api/serializers/__init__.py
@@ -27,3 +27,4 @@ from .machine_request_serializer import MachineRequestSerializer
 from .maintenance_record_serializer import MaintenanceRecordSerializer
 from .identity_detail_serializer import IdentityDetailSerializer
 from .atmo_user_serializer import AtmoUserSerializer
+from .profile_serializer import ProfileSerializer
